refactor(organon): extract energeia::planning submodule (#2642)
Move prographe and schedion tool definitions and executors to planning.rs.
